WIP on Shiny app
Still working on the shiny app. Currently struggling with the "csv" file names. There are many periods in the file name other than the last ".csv". This is what may need to be changed in order to have the shiny app be successful. Any help with this would be lovely